Factor out the UIA keyboard support into a mixin and add it to PuTTY/MinTTY
